requirements.txt
Django~=3.2
PyMySQL>=1.1.0
requests>=2.32.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
